+++ satella/coding/__init__.py
@@ -10,7 +10,7 @@ from .decorators import precondition, short_none, has_keys, \
 from .deleters import ListDeleter, DictDeleter
 from .fun_static import static_var
 from .iterators import hint_with_length, SelfClosingGenerator, exhaust, chain
-from .metaclasses import metaclass_maker, wrap_with, dont_wrap, wrap_property
+from .metaclasses import metaclass_maker, wrap_with, dont_wrap, wrap_property, DocsFromParent
 from .misc import update_if_not_none
 from .recast_exceptions import rethrow_as, silence_excs, catch_exception, log_exceptions, \
     raises_exception