+++ satella/coding/decorators/__init__.py
@@ -1,6 +1,7 @@
-from .arguments import auto_adapt_to_methods, attach_arguments, for_argument
+from .arguments import auto_adapt_to_methods, attach_arguments, for_argument, \
+    execute_before
 from .decorators import wraps, queue_get, chain_functions, has_keys, short_none
 from .preconditions import postcondition, precondition
 
-__all__ = ['postcondition', 'precondition', 'wraps', 'queue_get', 'chain_functions',
+__all__ = ['execute_before', 'postcondition', 'precondition', 'wraps', 'queue_get', 'chain_functions',
            'has_keys', 'short_none', 'auto_adapt_to_methods', 'attach_arguments', 'for_argument']