+++ satella/coding/misc.py
@@ -12,7 +12,8 @@ def update_if_not_none(dictionary: dict, key, value):
 
 
 def update_attr_if_none(obj: object, attr: str, value: tp.Any,
-                        on_attribute_error: bool = True) -> None:
+                        on_attribute_error: bool = True,
+                        if_value_is_not_none: bool = False) -> None:
     """
     Updates the object attribute, if it's value is None, or if
     it yields AttributeError (customizable as per on_attribute_error parameter)
@@ -23,16 +24,21 @@ def update_attr_if_none(obj: object, attr: str, value: tp.Any,
     :param on_attribute_error: whether to proceed with setting the value on
         AttributeError while trying to read given attribute. If False, AttributeError
         will be raised.
+    :param if_value_is_not_none: update object unconditionally, if only value is not None
     """
-    try:
-        val = getattr(obj, attr)
-        if val is None:
+    if if_value_is_not_none:
+        if value is not None:
             setattr(obj, attr, value)
-    except AttributeError:
-        if on_attribute_error:
-            setattr(obj, attr, value)
-        else:
-            raise
+    else:
+        try:
+            val = getattr(obj, attr)
+            if val is None:
+                setattr(obj, attr, value)
+        except AttributeError:
+            if on_attribute_error:
+                setattr(obj, attr, value)
+            else:
+                raise
 
 
 def update_key_if_none(dictionary: dict, key, value):