+++ CHANGELOG.md
@@ -1,5 +1,3 @@
 # v2.9.10
 
-
-
-
+* semantics of `catch_exception` changed