+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 # v2.9.10
 
 * semantics of `catch_exception` changed
+* added `DocsFromParent`
+