+++ CHANGELOG.md
@@ -1,2 +1,4 @@
 # v2.9.11
 
+* added `DocsFromParent` to __all__ in satella.coding
+