+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 # v2.9.11
 
 * added `DocsFromParent` to __all__ in satella.coding
-
+* added extra methods to `ListDeleter`, along with supporting reversed protocol