+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # v2.9.12
 
 * added `execute_before`
+* `PIDFileLock` will raise `RuntimeError` if `release` is called without `acquire`