+++ CHANGELOG.md
@@ -2,3 +2,6 @@
 
 * added `execute_before`
 * `PIDFileLock` will raise `RuntimeError` if `release` is called without `acquire`
+* added `returns_factory` to `silence_excs` and `rethrow_as`added `returns_factory` to 
+  `silence_excs` and `rethrow_as`
+