+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 # v2.9.13
 
+* `precondition` and `postcondition` are no longer debug-only
+