+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # v2.9.13
 
 * `precondition` and `postcondition` are no longer debug-only
+* added `CopyDocsFrom`