+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # v2.9.14
 
+* extra feature added to `update_attr_if_none`