+++ CHANGELOG.md
@@ -1,4 +1,2 @@
-# v2.9.14
-
-* extra feature added to `update_attr_if_none`
+# v2.9.15